requirements.txt
setuptools
python-dotenv
ipykernel
pandas
numpy
certifi
pymongo[srv]==3.11
pymongo
scikit-learn
-e .